# Speed up slicing graph generation