--- a/dask/array/core.py
+++ b/dask/array/core.py
@@ -45,7 +45,7 @@ from dask.array.dispatch import (  # noqa: F401
     tensordot_lookup,
 )
 from dask.array.numpy_compat import NUMPY_GE_200, _Recurser
-from dask.array.slicing import replace_ellipsis, setitem_array, slice_array
+from dask.array.slicing import SlicingNoop, replace_ellipsis, setitem_array, slice_array
 from dask.array.utils import (
     asanyarray_safe,
     asarray_safe,
@@ -2028,7 +2028,10 @@ class Array(DaskMethodsMixin):
             return self
 
         out = "getitem-" + tokenize(self, index2)
-        dsk, chunks = slice_array(out, self.name, self.chunks, index2, self.itemsize)
+        try:
+            dsk, chunks = slice_array(out, self.name, self.chunks, index2)
+        except SlicingNoop:
+            return self
 
         graph = HighLevelGraph.from_collections(out, dsk, dependencies=[self])