--- a/dask/array/slicing.py
+++ b/dask/array/slicing.py
@@ -97,7 +97,7 @@ def sanitize_index(ind):
         raise TypeError("Invalid index type", type(ind), ind)
 
 
-def slice_array(out_name, in_name, blockdims, index, itemsize):
+def slice_array(out_name, in_name, blockdims, index):
     """
     Main function for array slicing
 
@@ -118,8 +118,6 @@ def slice_array(out_name, in_name, blockdims, index, itemsize):
       This is the dask variable output name
     blockshape - iterable of integers
     index - iterable of integers, slices, lists, or None
-    itemsize : int
-        The number of bytes required for each element of the array.
 
     Returns
     -------
@@ -162,11 +160,7 @@ def slice_array(out_name, in_name, blockdims, index, itemsize):
     if all(
         isinstance(index, slice) and index == slice(None, None, None) for index in index
     ):
-        suffixes = product(*[range(len(bd)) for bd in blockdims])
-        dsk = {
-            (out_name,) + s: Alias((out_name,) + s, (in_name,) + s) for s in suffixes
-        }
-        return dsk, blockdims
+        raise SlicingNoop()
 
     # Add in missing colons at the end as needed.  x[5] -> x[5, :, :]
     not_none_count = sum(i is not None for i in index)
@@ -563,6 +557,12 @@ def issorted(seq):
     return np.all(seq[:-1] <= seq[1:])
 
 
+class SlicingNoop(Exception):
+    """This indicates that a slicing operation is a no-op. The caller has to handle this"""
+
+    pass
+
+
 def take(outname, inname, chunks, index, axis=0):
     """Index array with an iterable of index
 
@@ -581,9 +581,6 @@ def take(outname, inname, chunks, index, axis=0):
     >>> chunks
     ((4,),)
 
-    When any indexed blocks would otherwise grow larger than
-    dask.config.array.chunk-size, we will split them to avoid
-    growing chunksizes.
     """
 
     if not np.isnan(chunks[axis]).any():
@@ -592,14 +589,7 @@ def take(outname, inname, chunks, index, axis=0):
 
         arange = arange_safe(np.sum(chunks[axis]), like=index)
         if len(index) == len(arange) and np.abs(index - arange).sum() == 0:
-            # TODO: This should be a real no-op, but the call stack is
-            # too deep to do this efficiently for now
-            chunk_tuples = list(product(*(range(len(c)) for i, c in enumerate(chunks))))
-            graph = {
-                (outname,) + c: Alias((outname,) + c, (inname,) + c)
-                for c in chunk_tuples
-            }
-            return tuple(chunks), graph
+            raise SlicingNoop()
 
         average_chunk_size = int(sum(chunks[axis]) / len(chunks[axis]))
 
@@ -619,7 +609,7 @@ def take(outname, inname, chunks, index, axis=0):
         slices = [slice(None)] * len(chunks)
         slices[axis] = list(index)
         slices = tuple(slices)
-        chunk_tuples = list(product(*(range(len(c)) for i, c in enumerate(chunks))))
+        chunk_tuples = product(*(range(len(c)) for i, c in enumerate(chunks)))
         dsk = {
             (outname,)
             + ct: Task((outname,) + ct, getitem, TaskRef((inname,) + ct), slices)