This cleans up a couple of things around slicing to speed it up and restore it to its old performance. 

Closes https://github.com/dask/dask/issues/11735 (it's arguable whether it actually closes the ticket since some things are still a bit slower)

When comparing the old slicing performance with the new one, it is also important to note that the new slicing operation implements a shuffle for most nontrivial operations which is slower during graph generation but orders of magnitude better during runtime for most, if not all, affected operations.

There are many smaller performance fixes in here. However, the majority of speedup stems from disabling automatic garbage collection while this is running.


```

# Very old version
▶ python profile_slicing.py
Dask version: 2024.9.0
Empty indexer (all)                     :       0.01 ms
Every second element (::2)              :     146.57 ms
All but last element (:-1)              :     330.51 ms
Multi dim one slice (0, slice(None))    :      62.01 ms
Multi dim (0, slice(None, None, 2))     :      33.58 ms
Multi dim (0, slice(None, -1))          :      65.67 ms


# Main
(dask-distributed)
~/workspace/dask  ae250c9ac ✗                                                                                                                                                                                                         240d17h ◒
▶ python profile_slicing.py
Dask version: 2025.4.1+8.g4914f2f58
Empty indexer (all)                     :       0.01 ms
Every second element (::2)              :     467.75 ms
All but last element (:-1)              :     636.76 ms
Multi dim one slice (0, slice(None))    :     550.30 ms
Multi dim (0, slice(None, None, 2))     :     254.68 ms
Multi dim (0, slice(None, -1))          :     498.62 ms


# this PR
(dask-distributed)
~/workspace/dask  main ✗                                                                                                                                                                                                                 7d7h ◒
▶ python profile_slicing.py
Dask version: 2025.4.1+22.g2e6c80732
Empty indexer (all)                     :       0.02 ms
Every second element (::2)              :     156.41 ms
All but last element (:-1)              :     246.63 ms
Multi dim one slice (0, slice(None))    :     168.82 ms
Multi dim (0, slice(None, None, 2))     :      84.51 ms
Multi dim (0, slice(None, -1))          :     185.94 ms
```


<details>
<summary>Benchmark code</summary>

```python
import time
import dask.array as da
import dask

print(f"Dask version: {dask.__version__}")

x = da.ones((100_000,), chunks=(1,))
niterations = 5


def benchmark(arr, indexer, identifier):
    try:
        start = time.perf_counter()
        for i in range(niterations):
            arr[indexer]
        end = time.perf_counter()
        print(f"{identifier:<40}: {(end - start)/niterations * 1000:>10.2f} ms")
    except Exception as e:
        print(f"Error with {identifier}: {e}")
        return


benchmark(x, slice(None, None, None), "Empty indexer (all)")
benchmark(x, slice(None, None, 2), "Every second element (::2)")
benchmark(x, slice(None, -1), "All but last element (:-1)")

x = da.ones((100, 1000, 100), chunks=(-1, 1, 1))

benchmark(x, (1, slice(None)), "Multi dim one slice (0, slice(None))")
benchmark(x, (1, slice(None, None, 2)), "Multi dim (0, slice(None, None, 2))")
benchmark(x, (1, slice(None, -1)), "Multi dim (0, slice(None, -1))")
```

</details>