@fjetter 2025-05-12
TODO: Test if https://github.com/dask/dask/issues/11853 is fixed as well